Fix tip calculation
The tip calculation does not track with the comment. `BigInt(1e9) * BigInt(1e3)` is 1000 Gwei tip, not 100 Gwei tip. Either we want to give the proposer 1000 Gwei, in which case we should update the comment, or we we ant to give 100 in which case the code is wrong and has been fixed by this PR.